+++ rpcs3/Emu/SysCalls/Modules/cellResc.cpp
@@ -11,13 +11,13 @@
 Module *cellResc = nullptr;
 
 extern s32 cellVideoOutConfigure(u32 videoOut, vm::ptr<CellVideoOutConfiguration> config, vm::ptr<CellVideoOutOption> option, u32 waitForEvent);
-extern int cellGcmSetFlipMode(u32 mode);
+extern s32 cellGcmSetFlipMode(u32 mode);
 extern void cellGcmSetFlipHandler(vm::ptr<void(u32)> handler);
 extern void cellGcmSetVBlankHandler(vm::ptr<void(u32)> handler);
-extern int cellGcmAddressToOffset(u64 address, vm::ptr<be_t<u32>> offset);
-extern int cellGcmSetDisplayBuffer(u32 id, u32 offset, u32 pitch, u32 width, u32 height);
-extern int cellGcmSetPrepareFlip(vm::ptr<CellGcmContextData> ctx, u32 id);
-extern int cellGcmSetSecondVFrequency(u32 freq);
+extern s32 cellGcmAddressToOffset(u32 address, vm::ptr<be_t<u32>> offset);
+extern s32 cellGcmSetDisplayBuffer(u32 id, u32 offset, u32 pitch, u32 width, u32 height);
+extern s32 cellGcmSetPrepareFlip(vm::ptr<CellGcmContextData> ctx, u32 id);
+extern s32 cellGcmSetSecondVFrequency(u32 freq);
 extern u32 cellGcmGetLabelAddress(u8 index);
 extern u32 cellGcmGetTiledPitchSize(u32 size);
